requirements.txt
-f http://c.pypi.python.org/simple
BeautifulSoup==3.2.1
Coffin==0.3.8
Django==1.5.1
Jinja2==2.7.1
MarkupSafe==0.18
Pillow==2.1.0
PyYAML==3.10
Pygments==1.6
South==0.8.1
anyjson==0.3.3
django-appconf==0.6
django-choices==1.1.12
django-compressor==1.3
django-debug-toolbar==0.9.4
django-devserver==0.6.2
django-extensions==1.2.0
django-facebook==5.2.6
django-oauth2-provider==0.2.6
django-selectable==0.7.0
django-shell-ipynb==0.2.0
django-statsd-toolbar==0.1.2
django-tags-input==1.8.1
django-tastypie==0.9.15
docutils==0.10
dropbox2
eventlet==0.13.0
greenlet==0.4.1
html5lib==1.0b3
ipdb==0.7
ipython==1.0.0
lxml==3.2.1
mimeparse==0.1.3
ply==3.4
psycopg2==2.5.1
python-dateutil==2.1
python-utils
django-utils2
pyzmq==13.1.0
shortuuid==0.3
six==1.3.0
slimit==0.8.1
sorl-thumbnail==11.12
stevedore==0.11
tornado==3.1.1
uuid==1.30
virtualenv==1.10.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wsgiref==0.1.2
